Save the lastest SYSTEM_LAST_KMSG from dropbox when boot up
Dropbox generates a SYSTEM_LAST_KMSG file each time when the system reboot, power down, or crash.
And this information could be very useful for debug. So we check the dropbox and copy its latest SYSTEM_LAST_KMSG file
to the boot information folder. Check the history_event to find this folder and more information.

Original-Tracked-On: OAM-111964
Tracked-On: OAM-120395
Signed-off-by: Jade Guo <jade.guo@intel.com>